refactor(pipeline_scheduler): move PipelineScheduler logic mod.rs → scheduler.rs
mod.rs keeps only module docs, #![warn(missing_docs)], the launch/types
module decls + pub use types::*, and now mod scheduler; pub use scheduler::*.
All logic (PipelineScheduler + state/resolution types + impls) moved to
scheduler.rs verbatim. Zero behavior change (cargo check + clippy clean).
